// web/src/api/model/orgMemberResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * vaultctl API
 * Zero-knowledge, self-hosted password vault API
 * OpenAPI spec version: 1.0
 */

export interface OrgMemberResponse {
  acceptedAt?: string;
  invitedAt?: string;
  orgId?: string;
  role?: string;
  userId?: string;
}