// web/src/api/model/inviteResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * vaultctl API
 * Zero-knowledge, self-hosted password vault API
 * OpenAPI spec version: 1.0
 */

export interface InviteResponse {
  createdAt?: string;
  email?: string;
  expiresAt?: string;
  id?: string;
  inviterId?: string;
  role?: string;
}